refactor: options -> app_state
